Refactor storage internals into functional modules
