+++ packages/js-toolkit/utils/storage/createNoopProvider.ts
@@ -1,0 +1,18 @@
+import type { StorageProvider } from './types.js';
+
+export function createNoopProvider(): StorageProvider {
+  return {
+    get() {
+      return null;
+    },
+    set() {},
+    remove() {},
+    has() {
+      return false;
+    },
+    keys() {
+      return [];
+    },
+    clear() {},
+  };
+}
