+++ packages/js-toolkit/utils/storage/getGlobalStorage.ts
@@ -1,0 +1,5 @@
+export function getGlobalStorage(name: 'localStorage' | 'sessionStorage'): Storage | undefined {
+  return typeof globalThis !== 'undefined' && name in globalThis
+    ? globalThis[name]
+    : undefined;
+}
