+++ packages/js-toolkit/utils/storage/getStorageKeys.ts
@@ -1,0 +1,9 @@
+export function getStorageKeys(storage: Storage | undefined): string[] {
+  if (!storage) {
+    return [];
+  }
+
+  return Array.from({ length: storage.length }, (_, index) => storage.key(index)).filter(
+    (key): key is string => key !== null,
+  );
+}
